+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.3] - 2025-12-16
+
+### Added
+- Edit file description button in labeling toolbar navigation section
+- Reusable description dialog component for editing file descriptions
+- File description can now be edited directly from the labeling page
+
 ## [2.0.2] - 2025-11-03
 
 ### Added